refactor: ScrollObserver 수정
-  ScrollObserver 의 재사용성을 위한 수정
-  ScrollObserver 에서 APIHandlerForScroll , ChangedMovieListRenderer 분리
