+++ src/components/movie/MovieList.ts
@@ -1,4 +1,10 @@
-import { ScrollObserver } from '../../controller';
+import {
+  APIHandlerForScroll,
+  ChangedMovieListRenderer,
+  ScrollObserver,
+} from '../../controller';
+import { ListType, Movie } from '../../type/movie';
+import { createElementWithAttribute, ElementFinder } from '../../utils';
 
 import MovieItem from './MovieItem';
 import MovieListLastItem from './MovieListLastItem';
@@ -6,7 +12,6 @@ import NoneMovieItem from './NoneMovieItem';
 
 class MovieList {
   #element: HTMLElement;
-  #scrollObserver = new ScrollObserver();
   #isMovieList: boolean;
 
   constructor(movieList: Movie[] | undefined, isMoreData: boolean) {
@@ -45,15 +50,42 @@ class MovieList {
     return $ul;
   }
 
+  /**
+   * 영화 리스트 타입에 따라, 관련된 데이터를 불어와서 영화 리스트에 추가해 영화 리스트 ui를 변경한다.
+   * @param  {ListType} listType : 현재 보여지는 영화 리스트
+   */
+  async updateMovieData() {
+    const listType = this.#getListType();
+    if (!listType) return;
+    // 데이터 업데이트
+    await APIHandlerForScroll.handleGetMovieData(listType);
+    // 업데이트 된 영화 리스트 ui로 구현
+    ChangedMovieListRenderer.updateMovieList();
+  }
+
+  #getListType(): ListType | undefined {
+    const $movieListContainer = ElementFinder.findElementBySelector(
+      '.movie-list-container',
+    );
+
+    if (!$movieListContainer || !$movieListContainer.hasAttribute('name'))
+      return;
+
+    const name = $movieListContainer.getAttribute('name');
+
+    return name === 'search' ? 'search' : 'popular';
+  }
+
   #startObserving() {
     if (!this.#isMovieList) return;
 
-    const $scrollObserveTarget = this.#element.querySelector(
+    const $scrollObserverTarget = this.#element.querySelector(
       '.scroll-observer-target',
     );
-    if (!$scrollObserveTarget) return;
+    if (!$scrollObserverTarget) return;
 
-    this.#scrollObserver.observeTarget($scrollObserveTarget);
+    const scrollObserver = new ScrollObserver(this.updateMovieData.bind(this));
+    scrollObserver.observeTarget($scrollObserverTarget);
   }
 }
