+++ src/controller/index.ts
@@ -1,3 +1,5 @@
+import APIHandlerForScroll from './APIHandlerForScroll';
+import ChangedMovieListRenderer from './ChangedMovieListRenderer';
 import ScrollObserver from './ScrollObserver';
 
-export { ScrollObserver };
+export { APIHandlerForScroll, ChangedMovieListRenderer, ScrollObserver };
